RTC StormwaterModel - V2
In this version, we included post-processing of watershed modeling results. Moreover, we included the simulation of evapotranspiration and sub-surface exfiltration rate. In addition, we included two new optimization solvers for the MPC control (i.e., Global Search and Pattern Search). Furthermore, data now is input in .xlsx and .csv files, making the process easier. We also created a manual (not fully finished) to help use the model. Finally, we included the calculation of detention time to be able to perform flood and water quality control.